Fix pkg-config file to avoid overlinking
When configured to use external protobuf, remove -lprotobuf-lite from
pkg-config file's Libs section. The sentencepiece libraries link with
libprotobuf-lite; it is not necessary for users of the sentencepiece
libraries to link with protobuf unless they are using static linking so
this should have been in Libs.private. Use Requires.private so that we
use protobuf's pkg-config file to get the flags.